+++ examples/serdes.rs
@@ -8,9 +8,10 @@ extern crate bitvec;
 extern crate serde;
 extern crate serde_json;
 
+#[cfg(all(feature = "alloc", feature = "serde"))]
 use bitvec::prelude::*;
 
-#[cfg(all(feature = "alloc", feature = "serdes"))]
+#[cfg(all(feature = "alloc", feature = "serde"))]
 fn main() {
 	let bv = bitvec![1, 0, 1, 1, 0, 0, 1, 0];
 	let json = serde_json::to_string(&bv).expect("cannot fail to serialize");
@@ -21,7 +22,7 @@ fn main() {
 	assert_eq!(bb.as_slice()[0], 178);
 }
 
-#[cfg(not(all(feature = "alloc", feature = "serdes")))]
+#[cfg(not(all(feature = "alloc", feature = "serde")))]
 fn main() {
-	println!("This test needs to be compiled with `--features serdes` to work");
+	println!("This test needs to be compiled with `--features serde,std` to work");
 }